+++ AmazonChimeSDK/AmazonChimeSDK/analytics/EventAttributeName.swift
@@ -58,12 +58,14 @@ import Foundation
     case retryCount
     // Followings are related to Device Event Attributes
 
-    /// The error message that explains why the camera failed.
+    /// The error message that explains why the camera failed
     case videoInputError
-    /// The error message that explains why the microphone failed.
+    /// The error message that explains why the microphone failed
     case audioInputError
-    /// The error message that explains why the signaling dropped.
+    /// The error message that explains why the signaling websocket connection dropped
     case signalingDroppedError
+    /// The current app state
+    case appState
 
     public var description: String {
         switch self {
@@ -117,6 +119,8 @@ import Foundation
             return "videoInputError"
         case .signalingDroppedError:
             return "signalingDroppedError"
+        case .appState:
+            return "appState"
         }
     }
 }