+++ AmazonChimeSDK/AmazonChimeSDK/internal/audio/DefaultAudioClientObserver.swift
@@ -26,6 +26,7 @@ class DefaultAudioClientObserver: NSObject, AudioClientDelegate {
     private let logger: Logger
     private let eventAnalyticsController: EventAnalyticsController
     private let meetingStatsCollector: MeetingStatsCollector
+    private let appStateMonitor: AppStateMonitor
     private var primaryMeetingPromotionObserver: PrimaryMeetingPromotionObserver?
 
     private let audioLock: AudioLock
@@ -36,13 +37,15 @@ class DefaultAudioClientObserver: NSObject, AudioClientDelegate {
          configuration: MeetingSessionConfiguration,
          logger: Logger,
          eventAnalyticsController: EventAnalyticsController,
-         meetingStatsCollector: MeetingStatsCollector) {
+         meetingStatsCollector: MeetingStatsCollector,
+         appStateMonitor: AppStateMonitor) {
         self.audioClient = audioClient
         self.clientMetricsCollector = clientMetricsCollector
         self.configuration = configuration
         self.logger = logger
         self.eventAnalyticsController = eventAnalyticsController
         self.meetingStatsCollector = meetingStatsCollector
+        self.appStateMonitor = appStateMonitor
         audioLock = audioClientLock
         super.init()
         audioClient.delegate = self
@@ -410,6 +413,7 @@ class DefaultAudioClientObserver: NSObject, AudioClientDelegate {
     }
 
     private func handleAudioClientStop(status: MeetingSessionStatusCode) {
+        // TODO: only notifyAudioClientObserver() should be in the DispatchQueue?
         DispatchQueue.global().async {
             self.audioLock.lock()
             _ = self.audioClient.stopSession()
@@ -418,6 +422,7 @@ class DefaultAudioClientObserver: NSObject, AudioClientDelegate {
             self.notifyAudioClientObserver { (observer: AudioVideoObserver) in
                 observer.audioSessionDidStopWithStatus(sessionStatus: MeetingSessionStatus(statusCode: status))
             }
+            self.appStateMonitor.stop()
         }
     }