+++ AmazonChimeSDK/AmazonChimeSDK/session/DefaultMeetingSession.swift
@@ -29,19 +29,27 @@ import Foundation
         let clientMetricsCollector = DefaultClientMetricsCollector()
         let audioClientLock = NSLock()
 
-        let meetingStatsCollector =  DefaultMeetingStatsCollector(logger: logger)
+        let meetingStatsCollector = DefaultMeetingStatsCollector(logger: logger)
+        
+        let appStateMonitor: AppStateMonitor = DefaultAppStateMonitor(logger: self.logger)
 
-        self.eventAnalyticsController = DefaultEventAnalyticsController(meetingSessionConfig: configuration,
-                                                                        meetingStatsCollector: meetingStatsCollector,
-                                                                        logger: logger,
-                                                                        eventReporter: eventReporterFactory.createEventReporter())
+        let eventReporter: EventReporter? = eventReporterFactory.createEventReporter()
+        let defaultEventAnalyticsController = DefaultEventAnalyticsController(meetingSessionConfig: configuration,
+                                                                              meetingStatsCollector: meetingStatsCollector,
+                                                                              appStateMonitor: appStateMonitor,
+                                                                              logger: logger,
+                                                                              eventReporter: eventReporter)
+        appStateMonitor.delegate = defaultEventAnalyticsController
+        self.eventAnalyticsController = defaultEventAnalyticsController
+        
         let audioClientObserver = DefaultAudioClientObserver(audioClient: audioClient,
                                                              clientMetricsCollector: clientMetricsCollector,
                                                              audioClientLock: audioClientLock,
                                                              configuration: configuration,
                                                              logger: logger,
                                                              eventAnalyticsController: self.eventAnalyticsController,
-                                                             meetingStatsCollector: meetingStatsCollector)
+                                                             meetingStatsCollector: meetingStatsCollector,
+                                                             appStateMonitor: appStateMonitor)
         let activeSpeakerDetector =
             DefaultActiveSpeakerDetector(selfAttendeeId: configuration.credentials.attendeeId)
         let audioClientController = DefaultAudioClientController(audioClient: audioClient,
@@ -83,7 +91,7 @@ import Foundation
                                                                                          clientMetricsCollector: clientMetricsCollector,
                                                                                          eventAnalyticsController: eventAnalyticsController)
         let contentShareController = DefaultContentShareController(contentShareVideoClientController: contentShareVideoClientController)
-
+        
         self.audioVideo =
             DefaultAudioVideoFacade(audioVideoController:
                 DefaultAudioVideoController(audioClientController: audioClientController,
@@ -91,6 +99,7 @@ import Foundation
                                             clientMetricsCollector: clientMetricsCollector,
                                             videoClientController: videoClientController,
                                             videoTileController: videoTileController,
+                                            appStateMonitor: appStateMonitor,
                                             configuration: configuration,
                                             logger: logger),
                                     realtimeController: realtimeController,
@@ -117,6 +126,6 @@ import Foundation
         self.init(configuration: configuration,
                   logger: logger,
                   eventReporterFactory: DefaultMeetingEventReporterFactory(ingestionConfiguration: ingestionConfiguration,
-                                                                    logger: logger))
+                                                                           logger: logger))
     }
 }