+++ AmazonChimeSDK/AmazonChimeSDKTests/audiovideo/DefaultAudioVideoControllerTests.swift
@@ -19,6 +19,7 @@ class DefaultAudioVideoControllerTests: CommonTestCase {
     var clientMetricsCollectorMock: ClientMetricsCollectorMock!
     var videoClientControllerMock: VideoClientControllerMock!
     var videoTileControllerMock: VideoTileControllerMock!
+    var appStateMonitorMock: AppStateMonitorMock!
     var defaultAudioVideoController: DefaultAudioVideoController!
 
     override func setUp() {
@@ -29,12 +30,14 @@ class DefaultAudioVideoControllerTests: CommonTestCase {
         clientMetricsCollectorMock = mock(ClientMetricsCollector.self)
         videoClientControllerMock = mock(VideoClientController.self)
         videoTileControllerMock = mock(VideoTileController.self)
+        appStateMonitorMock = mock(AppStateMonitor.self)
 
         defaultAudioVideoController = DefaultAudioVideoController(audioClientController: audioClientControllerMock,
                                                                   audioClientObserver: audioClientObserverMock,
                                                                   clientMetricsCollector: clientMetricsCollectorMock,
                                                                   videoClientController: videoClientControllerMock,
                                                                   videoTileController: videoTileControllerMock,
+                                                                  appStateMonitor: appStateMonitorMock,
                                                                   configuration: meetingSessionConfigurationMock,
                                                                   logger: loggerMock)
     }
@@ -55,6 +58,7 @@ class DefaultAudioVideoControllerTests: CommonTestCase {
             reconnectTimeoutMs: self.reconnectTimeoutMs
         )).wasCalled()
         verify(videoClientControllerMock.start()).wasCalled()
+        verify(appStateMonitorMock.start()).wasCalled()
     }
 
     func testStart_callKitEnabled() {
@@ -213,6 +217,7 @@ class DefaultAudioVideoControllerTests: CommonTestCase {
 
         verify(audioClientControllerMock.stop()).wasCalled()
         verify(videoClientControllerMock.stopAndDestroy()).wasCalled()
+        verify(appStateMonitorMock.stop()).wasCalled()
     }
 
     func testAddAudioVideoObserver() {